Separate fs's unit tests into their own file
